api/requirements.txt
# Pinned 2026-04-29 — full freeze from running container for reproducible builds.
# Direct deps requested by this service:
#   fastapi[standard], pydantic, chromadb, langchain, langchain_openai, langchain_chroma
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
build==1.4.4
certifi==2026.4.22
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
fastapi==0.113.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.1
fastar==0.11.0
filelock==3.29.0
flatbuffers==25.12.19
fsspec==2026.3.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.12.2
idna==3.13
importlib_metadata==8.7.1
importlib_resources==7.1.0
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.15
langchain-chroma==1.1.0
langchain-core==1.3.2
langchain-openai==1.2.1
langchain-protocol==0.0.13
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.12
langgraph-sdk==0.3.13
langsmith==0.7.38
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.25.1
openai==2.33.0
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.2
protobuf==6.33.6
pybase64==1.4.3
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.23.1
tqdm==4.67.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.7.0
zipp==3.23.1
zstandard==0.25.0
